requirements.txt
flask==2.3.3
flask-cors==3.0.10
pillow==9.5.0
numpy==1.26.0
werkzeug==2.3.7
scipy==1.11.3
tensorflow==2.19.0
gunicorn
wget
